Added big changes to aula12's exercise. Details ahead :) :beach_umbrella:
Initially I thought I could use table names and column names as parameters
to the cursor.execute function, but turns out I can't.
See: https://stackoverflow.com/questions/35262753/python-mysqldb-variables-as-table-names

This was a big problem and got me thinking. With the help of that
stackoverflow question I figured the main reason we use parameterized
queries is to avoid SQL injection (very bad thing). In this particular
case, anyone can write/read to/from any table, so I ended just comparing
the value that comes from the user to a list of all of the tables in the
database. This list is one-time fetched in the start of the program.
This also applies to the columns.
The program will throw an Assertion Error that is treated in case it
finds that those values don't match.

Regarding the progress of this assignment, I still have alot to do.
The "getall()" function still needs to be done, but that is easy.
Then I need to code a kivy app that utilizes this API to provide a
frontend to an end user. Once I do that, I still need to complete
other easy exercises that I left behind due to those not having a
deadline. After that everything will be done and vacations start.
Can't wait.