How am I supposed to do the impossible?
I added a pdf file to exemplify what I wanted to do.
I have searched the coffins of the internet and I have found absolutely
nothing.........
It would be something as simple as a List thingy, but Kivy doesn't seem
to have it.....
Oh wait, it's deprecated... Like what?? It has been replaced by a so
called RecycleView. What is this name? How would I ever guess this??
smh

Imma try it now. I have wasted a whole afternoon with this.